aliexpress-api/aliexpressapi/requirements/dev.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2024.7.4
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.3.2
cryptography==45.0.7
Django==5.0.6
django-cors-headers==4.4.0
django-js-asset==2.2.0
django-mptt==0.16.0
django-nested-admin==4.1.4
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
factory-boy==3.3.0
Faker==26.0.0
idna==3.7
inflect==7.5.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
more-itertools==10.7.0
pillow==10.4.0
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-monkey-business==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
six==1.16.0
sqlparse==0.5.0
typeguard==4.4.4
typing_extensions==4.14.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uv==0.6.14
